+++ @navikt/core/react/src/modal/ModalContext.ts
@@ -1,0 +1,6 @@
+import React from "react";
+
+interface ModalContextProps {
+  closeHandler?: React.MouseEventHandler<HTMLButtonElement>;
+}
+export const ModalContext = React.createContext<ModalContextProps | null>(null);
